+++ openstack/bms/v2/tags/testing/requests_test.go
@@ -41,7 +41,7 @@ func TestCreateTag(t *testing.T) {
 	options := tags.CreateOpts{
 		Tag: []string{"__type_baremetal"},
 	}
-	n, err := tags.Create(fake.ServiceClient(), "2bff7a8a-3934-4f79-b1d6-53dc5540f00e", options).Extract()
+	n, err := tags.Create(fake.ServiceClient(), "2bff7a8a-3934-4f79-b1d6-53dc5540f00e", options)
 	th.AssertNoErr(t, err)
 	th.AssertEquals(t, "__type_baremetal", n.Tags[0])
 }
@@ -57,7 +57,7 @@ func TestDeleteTag(t *testing.T) {
 	})
 
 	res := tags.Delete(fake.ServiceClient(), "2bff7a8a-3934-4f79-b1d6-53dc5540f00e")
-	th.AssertNoErr(t, res.Err)
+	th.AssertNoErr(t, res)
 }
 
 func TestGetTags(t *testing.T) {
@@ -80,7 +80,7 @@ func TestGetTags(t *testing.T) {
 		`)
 	})
 
-	n, err := tags.Get(fake.ServiceClient(), "2bff7a8a-3934-4f79-b1d6-53dc5540f00e").Extract()
+	n, err := tags.Get(fake.ServiceClient(), "2bff7a8a-3934-4f79-b1d6-53dc5540f00e")
 	th.AssertNoErr(t, err)
 	th.AssertEquals(t, "__type_baremetal", n.Tags[0])
