--- openstack/bms/v2/tags/urls.go
@@ -1,9 +1,0 @@
-package tags
-
-import (
-	"github.com/opentelekomcloud/gophertelekomcloud"
-)
-
-func resourceURL(client *golangsdk.ServiceClient, serverId string) string {
-	return client.ServiceURL("servers", serverId, "tags")
-}
